requirements.txt
pydantic==1.8.2
PyYAML==6.0
typing_extensions==4.0.0
